+++ src/main/java/com/ftts/repository/RaceRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(path = "race")
 public interface RaceRepository extends JpaRepository<Race, Long>  {
-    Race getById(@Param("id") Integer id);
+    Race getById(@Param("id") Long id);
 
     @Query(value = "SELECT r from Race r WHERE raceName LIKE %:raceName%")
     Race getByRaceName(@Param("raceName") String raceName);
